Move Logic from POS to CashService
