+++ src/main/java/org/spin/pos/service/cash/CashServiceLogic.java
@@ -31,10 +31,15 @@ import org.adempiere.core.domains.models.I_AD_Process;
 import org.adempiere.core.domains.models.I_C_Payment;
 import org.adempiere.core.domains.models.X_C_Payment;
 import org.adempiere.exceptions.AdempiereException;
+import org.compiere.interfaces.PaymentProcessorClosing;
+import org.compiere.interfaces.PaymentProcessorStatus;
 import org.compiere.model.MBankAccount;
 import org.compiere.model.MBankStatement;
 import org.compiere.model.MPOS;
 import org.compiere.model.MPayment;
+import org.compiere.model.MPaymentProcessor;
+import org.compiere.model.PO;
+import org.compiere.model.PaymentProcessor;
 import org.compiere.model.Query;
 import org.compiere.process.DocAction;
 import org.compiere.util.DB;
@@ -48,6 +53,8 @@ import org.spin.backend.grpc.pos.CashClosingRequest;
 import org.spin.backend.grpc.pos.CashMovements;
 import org.spin.backend.grpc.pos.CashOpeningRequest;
 import org.spin.backend.grpc.pos.CashWithdrawalRequest;
+import org.spin.backend.grpc.pos.InfoOnlineCashClosingRequest;
+import org.spin.backend.grpc.pos.InfoOnlineCashClosingResponse;
 import org.spin.backend.grpc.pos.ListCashMovementsRequest;
 import org.spin.backend.grpc.pos.ListCashMovementsResponse;
 import org.spin.backend.grpc.pos.ListCashSummaryMovementsRequest;
@@ -57,6 +64,8 @@ import org.spin.backend.grpc.pos.PaymentSummary;
 import org.spin.backend.grpc.pos.PaymentTotal;
 import org.spin.backend.grpc.pos.PrintPreviewCashMovementsRequest;
 import org.spin.backend.grpc.pos.PrintPreviewCashMovementsResponse;
+import org.spin.backend.grpc.pos.ProcessOnlineCashClosingRequest;
+import org.spin.backend.grpc.pos.ProcessOnlineCashClosingResponse;
 import org.spin.backend.grpc.report_management.GenerateReportRequest;
 import org.spin.base.util.RecordUtil;
 import org.spin.grpc.service.ReportManagement;
@@ -600,6 +609,135 @@ public class CashServiceLogic {
 		;
 
 		return ticket;
+	}
+
+	public static ProcessOnlineCashClosingResponse.Builder processOnlineCashClosing(ProcessOnlineCashClosingRequest request) {
+		if (request.getId() <= 0) {
+			throw new AdempiereException("@FillMandatory@ @C_BankStatement_ID@");
+		}
+		int paymentMethodId = request.getPaymentMethodId();
+		if (paymentMethodId <= 0) {
+			throw new AdempiereException("@FillMandatory@ @C_PaymentMethod_ID@");
+		}
+		AtomicReference<ProcessOnlineCashClosingResponse.Builder> builderReference = new AtomicReference<>();
+		Trx.run( transactionName ->{
+			ProcessOnlineCashClosingResponse.Builder builder = ProcessOnlineCashClosingResponse.newBuilder();
+			MBankStatement bankStatement = new MBankStatement(Env.getCtx(), request.getId(), transactionName);
+			String whereClause = "C_BankAccount_ID = ?";
+			MPaymentProcessor paymentProcessor = (new Query(Env.getCtx(), "C_PaymentProcessor", whereClause, transactionName))
+					.setParameters(bankStatement.getC_BankAccount_ID())
+					.setClient_ID()
+					.setOnlyActiveRecords(true)
+					.first();
+			if (paymentProcessor == null) {
+				throw new AdempiereException("@C_PaymentProcessor_ID@ @NotFound@");
+			}
+			PO paymentProcessorRun = null;
+			try {
+				PaymentProcessor processor = PaymentProcessor.create(paymentProcessor, bankStatement, paymentMethodId);
+				if (PaymentProcessorClosing.class.isAssignableFrom(processor.getClass())) {
+					((PaymentProcessorClosing) processor).closeBatch(paymentMethodId, bankStatement.getStatementDate());
+					whereClause = "C_BankStatement_ID = ? AND C_PaymentMethod_ID = ?";
+					paymentProcessorRun = new Query(Env.getCtx(), "C_PaymentProcessorRun", whereClause, transactionName)
+							.setParameters(request.getId(), paymentMethodId)
+							.setOrderBy("Created DESC")
+							.first();
+
+				} else {
+					throw new AdempiereException(PaymentProcessorClosing.class.getName() + "Unsupported");
+				}
+			} catch (Exception e) {
+				throw new AdempiereException(e.getLocalizedMessage());
+			}
+			if (paymentProcessorRun == null) {
+				return;
+			}
+			String message = paymentProcessorRun.get_ValueAsString("ResponseMessage");
+			String status = paymentProcessorRun.get_ValueAsString("ResponseStatus");
+			boolean isError = "E".equals(status) || "R".equals(status);
+			builder
+					.setIsError(isError)
+					.setMessage(
+							StringManager.getValidString(
+									message
+							)
+					)
+					.setStatus(
+							StringManager.getValidString(
+									status
+							)
+					)
+					.setNextRequestTime(
+							paymentProcessorRun.get_ValueAsInt("NextRequestTime")
+					)
+			;
+			builderReference.set(builder);
+
+		});
+		return builderReference.get();
+	}
+
+	public static InfoOnlineCashClosingResponse.Builder infoOnlineCashClosing(InfoOnlineCashClosingRequest request) {
+		if (request.getId() <= 0) {
+			throw new AdempiereException("@FillMandatory@ @C_BankStatement_ID@");
+		}
+		int paymentMethodId = request.getPaymentMethodId();
+		if (paymentMethodId <= 0) {
+			throw new AdempiereException("@FillMandatory@ @C_PaymentMethod_ID@");
+		}
+		AtomicReference<InfoOnlineCashClosingResponse.Builder> builderReference = new AtomicReference<>();
+		Trx.run(transactionName -> {
+			InfoOnlineCashClosingResponse.Builder builder = InfoOnlineCashClosingResponse.newBuilder();
+			MBankStatement bankStatement = new MBankStatement(Env.getCtx(), request.getId(), transactionName);
+
+
+			String whereClause = "C_BankAccount_ID = ?";
+			MPaymentProcessor paymentProcessor = (new Query(Env.getCtx(), "C_PaymentProcessor", whereClause, transactionName))
+					.setParameters(bankStatement.getC_BankAccount_ID())
+					.setClient_ID()
+					.setOnlyActiveRecords(true)
+					.first();
+			if (paymentProcessor == null) {
+				throw new AdempiereException("@C_PaymentProcessor_ID@ @NotFound@");
+			}
+			PO paymentProcessorRun = null;
+			PaymentProcessor processor = PaymentProcessor.create(paymentProcessor, bankStatement, paymentMethodId);
+			if (PaymentProcessorStatus.class.isAssignableFrom(processor.getClass())) {
+				((PaymentProcessorStatus) processor).transactionStatus();
+				whereClause = "C_BankStatement_ID = ? AND C_PaymentMethod_ID = ?";
+				paymentProcessorRun = new Query(Env.getCtx(), "C_PaymentProcessorRun", whereClause, transactionName)
+						.setParameters(request.getId(), paymentMethodId)
+						.setOrderBy("Created DESC")
+						.first();
+
+			} else {
+				throw new AdempiereException(PaymentProcessorClosing.class.getName() + "Unsupported");
+			}
+			String	message = paymentProcessorRun.get_ValueAsString("ResponseMessage");
+			String	status = paymentProcessorRun.get_ValueAsString("ResponseStatus");
+
+
+			boolean isError = "E".equals(status) || "R".equals(status);
+			builder
+					.setIsError(isError)
+					.setMessage(
+							StringManager.getValidString(
+									message
+							)
+					)
+					.setStatus(
+							StringManager.getValidString(
+									status
+							)
+					)
+					.setNextRequestTime(
+							paymentProcessorRun.get_ValueAsInt("NextRequestTime")
+					)
+			;
+			;
+			builderReference.set(builder);
+		});
+		return builderReference.get();
 	}
 
 }
